aliyun_server/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-cors==4.0.0
pymysql==1.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
redis==5.0.1
celery==5.3.4
aliyun-python-sdk-core==2.14.0
oss2==2.18.4
loguru==0.7.2
